Vectorized SequenceCompareTo for Span<char>
